requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
anaconda-client==1.11.1
anaconda-project==0.11.1
annotated-types==0.7.0
anyio==4.4.0
archspec==0.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
backports.functools-lru-cache==1.6.4
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
bokeh==3.4.3
boltons==23.0.0
Brotli==1.0.9
cachetools==5.4.0
catalogue==1.0.2
certifi==2024.6.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.0.0
clyent==1.2.2
colorcet==3.1.0
conda-content-trust==0.1.3
conda-pack==0.6.0
conda-package-handling==2.3.0
conda_package_streaming==0.10.0
conda-repo-cli==1.0.88
conda-verify==3.4.2
confection==0.1.5
contourpy==1.2.1
cryptography==42.0.5
cucim-cu12==24.4.0
cuda-python==12.6.0
cudf-cu12==24.4.1
cugraph-cu12==24.4.0
cuml-cu12==24.4.0
cuproj-cu12==24.4.0
cupy-cuda12x==13.2.0
cuspatial-cu12==24.4.0
cuvs-cu12==24.4.0
cuxfilter-cu12==24.4.1
cymem==2.0.8
dask==2024.1.1
dask-cuda==24.4.0
dask-cudf-cu12==24.4.1
dask-expr==0.4.0
datasets==2.19.2
datashader==0.16.3
defusedxml==0.7.1
dill==0.3.8
distributed==2024.1.1
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
en-vectors-web-lg==2.1.0
exceptiongroup==1.2.2
fastjsonschema==2.16.2
fastrlock==0.8.2
filelock==3.13.1
fqdn==1.5.1
frozendict==2.4.2
frozenlist==1.4.1
fsspec==2024.3.1
future==0.18.3
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
h5py==3.11.0
holoviews==1.19.1
huggingface-hub==0.23.4
idna==3.7
imageio==2.35.0
importlib_metadata==8.2.0
inquirerpy==0.3.4
isoduration==20.11.0
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.2
jupyter_core==5.5.0
jupyter-events==0.10.0
jupyter_server==2.14.2
jupyter_server_proxy==4.3.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
libarchive-c==2.9
libmambapy==1.5.8
lightning==2.4.0
lightning-utilities==0.11.6
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdit-py-plugins==0.4.1
mdurl==0.1.2
menuinst==2.1.1
mistune==3.0.2
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.16
murmurhash==1.0.10
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.9.2
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
nvtx==0.2.10
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
panel==1.4.5
param==2.1.1
partd==1.4.2
pfzy==0.3.4
pillow==10.3.0
pip==22.3.1
pkginfo==1.10.0
plac==1.1.3
platformdirs==3.10.0
pluggy==1.0.0
ply==3.11
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pycosat==0.6.6
pycparser==2.21
pyct==0.5.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pylibcugraph-cu12==24.4.0
pylibraft-cu12==24.4.0
pynvjitlink-cu12==0.3.0
pynvml==11.4.1
pyogrio==0.9.0
pyproj==3.6.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-lightning==2.4.0
pytz==2024.1
pyviz_comms==3.0.3
PyYAML==6.0.1
pyzmq==26.1.0
QtPy==2.4.1
raft-dask-cu12==24.4.0
rapids-dask-dependency==24.4.1
referencing==0.30.2
regex==2024.7.24
requests==2.32.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rmm-cu12==24.4.0
rpds-py==0.10.6
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruamel-yaml-conda==0.17.21
safetensors==0.4.4
scikit-image==0.22.0
scikit-learn==1.5.1
scipy==1.14.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.12.0
setproctitle==1.3.3
setuptools==65.6.3
shapely==2.0.5
shellingham==1.5.4
simpervisor==1.0.0
sip==6.7.12
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==2.3.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==1.0.7
sympy==1.13.2
tblib==3.0.0
tensorboardX==2.6.2.2
terminado==0.18.1
thinc==7.4.6
threadpoolctl==3.5.0
tifffile==2024.8.10
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.1
torch==2.4.0
torchaudio==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tornado==6.3.3
tqdm==4.66.4
traitlets==5.14.3
transformers==4.44.0
treelite==4.1.2
triton==3.0.0
truststore==0.8.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
ucx-py-cu12==0.37.0
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.1
wandb==0.17.6
wasabi==0.10.1
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.38.4
wrapt==1.16.0
xarray==2024.7.0
xxhash==3.4.1
xyzservices==2024.6.0
yarl==1.9.4
zict==3.0.0
zipp==3.20.0
zstandard==0.22.0
